Add logic for calculating intervals
The key is determining the start of the interval, then working from there. PHP's `DateTimeInterface` handles a lot of this, but calculating the start of the week based on the current day of the week *and* respecting the 'week_starts_on' option requires some math.

For the sake of testing, the `OrderLimiter` constructor now accepts a `DateTimeImmutable` representing "now", which all further calculations will be based on (default is `current_datetime()`).