Use WordPress 5.3 in the "Minimum Requirements" CI job
The plugin headers declare WordPress 5.3 as the minimum supported version, yet the "Minimum requirements" Travis CI job was using WordPress 5.2.

The reason we use 5.3 is for [the date/time improvements](https://make.wordpress.org/core/2019/09/23/date-time-improvements-wp-5-3/), as well as the fact that 5.2 is old (current WP release branch is 5.4.x); there's no sense in polyfilling this functionality as store owners *should* be running the latest version of WordPress.